Clean up INFO logs and fix ConstraintSet warnings
- Remove verbose INFO-level logging from multiple components:
  - TorPortManager, TorPluginFactory, TorPlugin
  - AndroidTaskScheduler, AndroidNetworkManager
  - CleanupManagerImpl, ConnectionRegistryImpl
  - PollerImpl, DuplexOutgoingSession
  - PluginManagerImpl, LifecycleManagerImpl
  - AccountManagerImpl, H2Database, JdbcDatabase

- Fix ConstraintSet warnings for replyPreviewContainer and voiceMessageView
  - Add placeholder views in list_item_conversation_msg_image.xml
  - Add placeholder views in list_item_conversation_msg_image_text.xml

- Include previous video attachment and chunked attachment improvements

+++ briar-core/src/main/java/org/briarproject/briar/avatar/AvatarValidator.java
@@ -89,7 +89,7 @@ class AvatarValidator implements MessageValidator {
 		BdfDictionary meta = new BdfDictionary();
 		meta.put(MSG_KEY_VERSION, version);
 		meta.put(MSG_KEY_CONTENT_TYPE, contentType);
-		meta.put(MSG_KEY_DESCRIPTOR_LENGTH, descriptorLength);
+		meta.put(MSG_KEY_DESCRIPTOR_LENGTH, (int) descriptorLength);
 		return meta;
 	}
